// src/markdown/types.ts
export interface HtmlResults {
  /**
   * Anchor (link) data and in order found in the document. Does not include
   * anchors without `href` attributes or href's that start with `#`.
   */
  anchors: AnchorData[];
  /**
   * Results of parsing the html into a JSON serializable format which can
   * be later used to generate JSX/hypertext, both serverside and clientside.
   */
  ast: JsxAstNode[];
  /**
   * Heading data and in order found in the document.
   */
  headings: HeadingData[];
  /**
   * Images in order found in the document.
   */
  imgs: ImgData[];
  /**
   * All the HTML tags found, but no duplicates.
   */
  tagNames: string[];
  /**
   * The resulting HTML, which may be different from the passed in
   * HTML due to any changes that could have happened within the
   * `beforeHtmlSerialize(frag)` option.
   */
  html: string;
}

export interface MarkdownResults<T = { [key: string]: string }>
  extends HtmlResults {
  /**
   * Contains extracted yaml attributes.
   */
  attributes: T;
  /**
   * The description from the front matter attributes.
   */
  description?: string;
  /**
   * The resolved file path of the markdown file.
   */
  filePath?: string;
  /**
   * Results from parsing the markdown into html.
   */
  html: string;
  /**
   * Slugified title from the filename provided in the parse options. This can be
   * overridden by adding a `slug` front matter attribute.
   */
  slug?: string;
  /**
   * The title from the front matter attributes.
   */
  title?: string;
}

export interface PageUrlOptions {
  /**
   * If the url should always end with a `/` or not. Default is to not end with a trailing `/`.
   * @default false
   */
  trailingSlash?: boolean;
}

export interface PageNavigationOptions extends PageUrlOptions {
  /**
   * The table of contents which can be used to calculate the previous and next pages.
   */
  tableOfContents?: TableOfContents;
}

export interface PageNavigation {
  current: PageNavigationData | null;
  parent: PageNavigationData | null;
  next: PageNavigationData | null;
  previous: PageNavigationData | null;
}

export interface PageNavigationData {
  title: string | null;
  url: string | null;
}

export interface AnchorData {
  href: string | null;
  text: string;
}

export interface HeadingData {
  text: string;
  level: number;
  id: string | null;
}

export interface ImgData {
  text: string | null;
  src: string | null;
}

export type JsxAstNode = any;
// ...
export interface TableOfContents {
  tocFilePath: string;
  tocDirPath: string;
  rootPagesDir: string;
  root: TableOfContentsNode[];
}

export interface TableOfContentsNode {
  text?: string;
  url?: string;
  file?: string;
  hasParent?: boolean;
  children?: TableOfContentsNode[];
  depth: number;
}